requirements.txt
beautifulsoup4==4.7.1
boto3==1.4.7
botocore==1.7.43
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.7
django-allauth==0.38.0
django-contrib-comments==1.9.1
django-dropbox-storage==0.4.1
django-heroku==0.3.1
django-report-tools==0.2.2
django-sesame==1.4
django-social-share==1.3.2
django-storages==1.7.1
django-tagging==0.4.6
django-widget-tweaks==1.4.3
docutils==0.14
dropbox==9.3.0
Faker==1.0.4
gunicorn==19.9.0
html5lib==0.999999999
idna==2.8
jmespath==0.9.3
markdown==2.6.9
mysqlclient==1.3.13
nltk==3.4
numpy==1.16.2
oauthlib==2.1.0
pbr==4.0.2
Pillow==5.3.0
psycopg2==2.7.6.1
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.7
requests==2.21.0
requests-oauthlib==1.0.0
s3transfer==0.1.11
singledispatch==3.4.0.3
six==1.10.0
soupsieve==1.8
stevedore==1.28.0
text-unidecode==1.2
urllib3==1.24.1
virtualenv==15.1.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
webencodings==0.5.1
whitenoise==4.1.2
